Bug #990075 misc: sprdwcn: modify output formats of int to support ums510_1h10_go
[root cause  ] modify output formats of int
[changes     ] modify output formats of int
[side effects] no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:no
[is there dependence]:no
[confirm dependent commit]:no
[board]:ums510
[test case]:it compiles okay
[reviewers   ] Jingxiang.li
[change_type ] bringup --> sprd_chip_support
[tag_product ] common
Test: it compiles okay.

The patch modifies output formats of int to support ums510_1h10_go.

+++ drivers/misc/sprdwcn/platform/wcn_boot.c
@@ -758,7 +758,8 @@ static int marlin_parse_dt(struct platform_device *pdev)
 		WCN_ERR("gpio_rst request err: %d\n",
 				marlin_dev->int_ap);
 
-	WCN_INFO("BTWF_FIRMWARE_PATH len=%ld\n", strlen(BTWF_FIRMWARE_PATH));
+	WCN_INFO("BTWF_FIRMWARE_PATH len=%ld\n",
+		 (long)strlen(BTWF_FIRMWARE_PATH));
 	ret = of_property_read_string(np, "sprd,btwf-file-name",
 				      (const char **)&marlin_dev->btwf_path);
 	if (!ret) {
@@ -767,7 +768,8 @@ static int marlin_parse_dt(struct platform_device *pdev)
 		WCN_INFO("BTWG path is %s\n", BTWF_FIRMWARE_PATH);
 	}
 
-	WCN_INFO("BTWF_FIRMWARE_PATH2 len=%ld\n", strlen(BTWF_FIRMWARE_PATH));
+	WCN_INFO("BTWF_FIRMWARE_PATH2 len=%ld\n",
+		 (long)strlen(BTWF_FIRMWARE_PATH));
 
 	ret = of_property_read_string(np, "sprd,gnss-file-name",
 				      (const char **)&marlin_dev->gnss_path);
